moved tests + removed tabs state
